Add ssl certificate api. Code cleanup. File organized to indicate module from which there are.
